#67 Convert authorization and authentication services into DI classes
