+++ src/service.locator.ts
@@ -1,0 +1,6 @@
+import { AuthenticationService } from './auth/authentication.service';
+import { AuthorisationService } from './auth/authorisation.service';
+
+// auth
+export const authenticationService = new AuthenticationService();
+export const authorisationService = new AuthorisationService();
